feat(commands): split admin commands
